Restructuring classes to avoid duplicated implicit conversions
