+++ airframe-rx-html/.js/src/main/scala/wvlet/airframe/rx/html/compat.scala
@@ -18,7 +18,7 @@ import org.scalajs.dom
   */
 object compat {
   trait PlatformEmbeddableNode {
-    @inline implicit def embedHtmlElement[A <: dom.Element]: EmbeddableNode[A] = null
+    @inline implicit def embedHtmlElement[A <: dom.Element]: RxEmbedding.EmbeddableNode[A] = null
   }
   type UIEvent    = dom.UIEvent
   type MouseEvent = dom.MouseEvent
