+++ airframe-rx-html/src/main/scala/wvlet/airframe/rx/html/RxElement.scala
@@ -15,6 +15,7 @@ package wvlet.airframe.rx.html
 
 import wvlet.airframe.rx.{Cancelable, Rx}
 import wvlet.log.LogSupport
+import wvlet.airframe.rx.html.RxEmbedding._
 
 /**
   */
